+++ addon/query/predicate.js
@@ -293,25 +293,25 @@ export class StringPredicate extends BasePredicate {
 }
 
 /**
- * The predicate class for geography attributes.
+ * The base class of logical predicate for spatial attributes.
  *
  * @namespace Query
- * @class GeographyPredicate
+ * @class SpatialPredicate
  * @extends BasePredicate
  *
  * @param {String} attributePath The path to the attribute for predicate.
  * @constructor
  */
-export class GeographyPredicate extends BasePredicate {
+export class SpatialPredicate extends BasePredicate {
   constructor(attributePath) {
     super();
 
     if (!attributePath) {
-      throw new Error('Attribute path is required for GeographyPredicate constructor.');
+      throw new Error('Attribute path is required for SpatialPredicate constructor.');
     }
 
     this._attributePath = attributePath;
-    this._intersectsValue = null;
+    this._options = null;
   }
 
   /**
@@ -326,27 +326,141 @@ export class GeographyPredicate extends BasePredicate {
   }
 
   /**
-   * The geography value that has to intersect with the attribute.
+   * The filter operator.
    *
-   * @property intersectsValue
-   * @type {String}
+   * @property operator
+   * @type Query.FilterOperator
    * @public
    */
-  get intersectsValue() {
-    return this._intersectsValue;
+   get operator() {
+    return this._options.operator;
   }
 
   /**
-   * Sets the value that the attribute has to contain.
+   * The spatial type value that has to be the second argument of the spatial function,
+   * assuming the attribute is the first argument.
    *
-   * @method contains
-   * @param {String} geography The geography value that has to intersect with the attribute.
-   * @return {Query.StringPredicate} Returns this instance.
+   * @property spatial
+   * @type {String}
+   * @public
+   */
+   get spatial() {
+    return this._options.spatial;
+  }
+
+  /**
+   * The spatial function to call.
+   *
+   * @property spatialFunction
+   * @type {String}
+   * @public
+   */
+  get spatialFunction() {
+    return this._function;
+  }
+
+  /**
+   * The spatial type namespace.
+   *
+   * @property spatialNamespace
+   * @type {String}
+   * @public
+   */
+  get spatialNamespace() {
+      throw this._getAbstractPropertyError('spatialNamespace');
+  }
+
+  /**
+   * The spatial type.
+   *
+   * @property spatialType
+   * @type {String}
+   * @public
+   */
+  get spatialType() {
+      throw this._getAbstractPropertyError('spatialType');
+  }
+
+  /**
+   * The value for filtering.
+   *
+   * @property value
+   * @type String|Number
+   * @public
+   */
+   get value() {
+    return this._options.value;
+  }
+
+  /**
+   * Sets the spatial predicate metadata that has to be used on build the distance condition with the attribute.
+   * 
+   * @method distance
+   * @param {String} spatial The spatial type value that has to have a distance from the attribute.
+   * @param operator {Query.FilterOperator|String} The filter operator.
+   * @param value {String|Number} The value for filtering.
+   * @return {Query.SpatialPredicate} Returns this instance.
    * @chainable
    */
-  intersects(geography) {
-    this._intersectsValue = geography;
+  distance(spatial, operator, value) {
+    this._function = 'distance';
+    this._options = {
+      spatial: spatial,
+      operator: FilterOperator.tryCreate(operator),
+      value: value,
+    }
     return this;
+  }
+
+  /**
+   * Sets the spatial predicate metadata that has to be used on build the intersection condition with the attribute.
+   *
+   * @method intesects
+   * @param {String} spatial The spatial type value that has to intersect with the attribute.
+   * @return {Query.SpatialPredicate} Returns this instance.
+   * @chainable
+   */
+  intersects(spatial) {
+    this._function = 'intersects';
+    this._options = {
+      spatial: spatial,
+    }
+    return this;
+  }
+
+  _getAbstractPropertyError(property) {
+    return new Error(`The ${property} property must be overridden in the SpatialPredicate subtypes.`);
+  }
+}
+
+/**
+ * The predicate class for geography attributes.
+ *
+ * @namespace Query
+ * @class GeographyPredicate
+ * @extends SpatialPredicate
+ */
+export class GeographyPredicate extends SpatialPredicate {
+  /**
+   * The geography type namespace.
+   *
+   * @property spatialNamespace
+   * @type {String}
+   * @public
+   */
+  get spatialNamespace() {
+    return 'geo';
+  }
+
+  /**
+   * The geography type.
+   *
+   * @property spatialType
+   * @type {String}
+   * @public
+   */
+  get spatialType() {
+    return 'geography';
   }
 }
 
@@ -355,56 +469,28 @@ export class GeographyPredicate extends BasePredicate {
  *
  * @namespace Query
  * @class GeometryPredicate
- * @extends BasePredicate
- *
- * @param {String} attributePath The path to the attribute for predicate.
- * @constructor
+ * @extends SpatialPredicate
  */
-export class GeometryPredicate extends BasePredicate {
-  constructor(attributePath) {
-    super();
-
-    if (!attributePath) {
-      throw new Error('Attribute path is required for GeometryPredicate constructor.');
-    }
-
-    this._attributePath = attributePath;
-    this._intersectsValue = null;
-  }
-
-  /**
-   * The path to the attribute for predicate.
+export class GeometryPredicate extends SpatialPredicate {  /**
+   * The geometry type namespace.
    *
-   * @property attributePath
+   * @property spatialNamespace
    * @type {String}
    * @public
    */
-  get attributePath() {
-    return this._attributePath;
+  get spatialNamespace() {
+    return 'geom';
   }
 
   /**
-   * The geometry value that has to intersect with the attribute.
+   * The geometry type.
    *
-   * @property intersectsValue
+   * @property spatialType
    * @type {String}
    * @public
    */
-  get intersectsValue() {
-    return this._intersectsValue;
-  }
-
-  /**
-   * Sets the value that the attribute has to contain.
-   *
-   * @method contains
-   * @param {String} geometry The geometry value that has to intersect with the attribute.
-   * @return {Query.StringPredicate} Returns this instance.
-   * @chainable
-   */
-  intersects(geometry) {
-    this._intersectsValue = geometry;
-    return this;
+  get spatialType() {
+    return 'geometry';
   }
 }